requirements.txt
fastapi
uvicorn
python-multipart
praat-parselmouth
pedalboard
numpy
scipy
